+++ README.md
@@ -1,1 +1,4 @@
 My first ready to use bot
+
+This bot uses webhooks for connection
+with heroku server, where he deployed
